[KOGITO-9056] Refactoring executor to conditionally load handlers
