+++ kogito-serverless-workflow/kogito-serverless-workflow-executor-core/src/main/resources/META-INF/services/org.kie.kogito.serverless.workflow.executor.StaticWorkflowRegister
@@ -1,0 +1,1 @@
+org.kie.kogito.serverless.workflow.executor.StaticJavaRegister
