+++ kogito-serverless-workflow/kogito-serverless-workflow-executor-core/src/main/resources/META-INF/services/org.kie.kogito.serverless.workflow.parser.FunctionTypeHandler
