+++ kogito-serverless-workflow/kogito-serverless-workflow-executor-core/src/test/resources/interpolation.sw.json
