+++ kogito-serverless-workflow/kogito-serverless-workflow-executor-rest/src/main/java/org/kie/kogito/serverless/workflow/executor/StaticRestRegister.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2023 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.serverless.workflow.executor;
+
+import io.vertx.mutiny.core.Vertx;
+
+public class StaticRestRegister implements StaticApplicationRegister {
+
+    private final Vertx vertx;
+    private final StaticRestWorkItemHandler restWorkItemHandler;
+
+    public StaticRestRegister() {
+        vertx = Vertx.vertx();
+        restWorkItemHandler = new StaticRestWorkItemHandler(vertx);
+    }
+
+    @Override
+    public void register(StaticWorkflowApplication application) {
+        application.registerHandler(restWorkItemHandler);
+    }
+
+    @Override
+    public void close() {
+        restWorkItemHandler.close();
+        vertx.closeAndAwait();
+    }
+}
