+++ kogito-serverless-workflow/kogito-serverless-workflow-executor-rest/src/main/java/org/kie/kogito/serverless/workflow/executor/StaticRestWorkItemHandler.java
