+++ kogito-serverless-workflow/kogito-serverless-workflow-executor-rest/src/main/resources/META-INF/services/org.kie.kogito.serverless.workflow.executor.StaticApplicationRegister
@@ -1,0 +1,1 @@
+org.kie.kogito.serverless.workflow.executor.StaticRestRegister
