+++ kogito-serverless-workflow/kogito-serverless-workflow-executor-rest/src/test/resources/interpolation.sw.json
@@ -1,0 +1,30 @@
+{  
+  "id": "interpolation",
+  "version": "1.0",
+  "name": "Interpolation",
+  "description": "Sample interpolation",
+  "start": "Hello State",
+  "functions": [ 
+    { 
+      "name": "logInfo",
+      "type": "custom",
+      "operation": "sysout:INFO"
+    }
+  ]
+  ,
+  "states":[  
+    {  
+     "name":"Hello State",
+     "type":"operation",
+     "actions": [{
+     	"functionRef" : {
+          "refName": "logInfo",
+          "arguments" : {
+            "message": "\"Model is \\(.)\""
+          }
+        }
+     }],
+     "end": true
+   }
+  ]
+}
