+++ kogito-serverless-workflow/kogito-serverless-workflow-executor/pom.xml
@@ -11,41 +11,21 @@
 
   <artifactId>kogito-serverless-workflow-executor</artifactId>
   <name>Kogito :: Serverless Workflow :: Executor</name>
-  <description>Classes for execution of a workflow without code generation</description>
+  <description>Dependencies for execution of a workflow without code generation</description>
+  <packaging>pom</packaging>
 
   <dependencies>
+     <dependency>
+       <groupId>org.kie.kogito</groupId>
+       <artifactId>kogito-serverless-workflow-executor-core</artifactId>
+    </dependency>
+     <dependency>
+       <groupId>org.kie.kogito</groupId>
+       <artifactId>kogito-serverless-workflow-executor-rest</artifactId>
+    </dependency>
     <dependency>
        <groupId>org.kie.kogito</groupId>
-       <artifactId>kogito-serverless-workflow-builder</artifactId>
-    </dependency>
-      <dependency>
-       <groupId>org.kie.kogito</groupId>
-       <artifactId>kogito-serverless-workflow-rest-parser</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter-engine</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter-params</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>ch.qos.logback</groupId>
-      <artifactId>logback-classic</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.github.tomakehurst</groupId>
-      <artifactId>wiremock-jre8</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.assertj</groupId>
-      <artifactId>assertj-core</artifactId>
-      <scope>test</scope>
+       <artifactId>kogito-serverless-workflow-fluent</artifactId>
     </dependency>
    </dependencies>
 </project>
