--- kogito-serverless-workflow/kogito-serverless-workflow-executor/src/main/java/org/kie/kogito/serverless/workflow/executor/JavaFunctionExtension.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2023 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.serverless.workflow.executor;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-
-import io.serverlessworkflow.api.interfaces.Extension;
-
-public class JavaFunctionExtension implements Extension {
-
-    public static final String EXTENSION_ID = "java_functions";
-
-    @Override
-    public String getExtensionId() {
-        return EXTENSION_ID;
-    }
-
-    private Map<String, Function<?, ?>> functions = new HashMap<>();
-
-    public Map<String, Function<?, ?>> functions() {
-        return functions;
-    }
-}
