+++ kogito-serverless-workflow/kogito-serverless-workflow-fluent/pom.xml
@@ -1,0 +1,22 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.kie.kogito</groupId>
+    <artifactId>kogito-serverless-workflow</artifactId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>kogito-serverless-workflow-fluent</artifactId>
+  <name>Kogito :: Serverless Workflow :: Fluent</name>
+  <description>Classes for fluent definition of a serverless workflow</description>
+
+  <dependencies>
+    <dependency>
+       <groupId>org.kie.kogito</groupId>
+       <artifactId>kogito-serverless-workflow-builder</artifactId>
+    </dependency>
+   </dependencies>
+</project>
