+++ kogito-serverless-workflow/pom.xml
@@ -25,6 +25,9 @@
     <module>kogito-serverless-workflow-rest-runtime</module>
     <module>kogito-serverless-workflow-runtime</module>
     <module>kogito-serverless-workflow-executor</module>
+    <module>kogito-serverless-workflow-executor-core</module>
+    <module>kogito-serverless-workflow-executor-rest</module>
+    <module>kogito-serverless-workflow-fluent</module>
     <module>kogito-serverless-workflow-utils</module>
     <module>kogito-jsonpath-expression</module>
     <module>kogito-jq-expression</module>
